requirements.txt
google-cloud-storage>=2.10.0
google-cloud-resource-manager>=1.10.0
google-auth>=2.22.0
pyyaml>=6.0.1
tenacity>=8.2.3
aiohttp>=3.8.5
